requirements.txt
Package             Version
------------------- -----------
aniso8601           9.0.1
blinker             1.7.0
click               8.1.7
colorama            0.4.6
dnspython           2.5.0
email-validator     2.1.0.post1
Flask               3.0.2
Flask-Login         0.6.3
Flask-Principal     0.4.0
Flask-RESTful       0.3.10
Flask-Security-Too  5.3.3
Flask-SQLAlchemy    3.1.1
Flask-WTF           1.2.1
greenlet            3.0.3
idna                3.6
importlib-resources 6.1.1
itsdangerous        2.1.2
Jinja2              3.1.2
MarkupSafe          2.1.5
passlib             1.7.4
pip                 21.2.4
pytz                2024.1
setuptools          58.1.0
six                 1.16.0
SQLAlchemy          2.0.25
typing_extensions   4.9.0
Werkzeug            3.0.1
WTForms             3.1.2